tidy code and comments

+++ listEditable.php
@@ -19,19 +19,25 @@ function listEditable() : String {
     $result = $query->fetchAll();
 
     $return = '';
-
+    /**
+     * If a title is set for a game, the relevant section of the page should have en entry added to it with a form
+     * allowing that game's information to be edited or the game to be removed from the database (since nothing confidential
+     * is being stored, this database uses a soft deletion)
+     */
     foreach ($result as $game) {
-        $return .=     '<div class=editGame>'
-                         . '<form method="get" action="./editEntry.php">'
-                             . '<input class="hidden" type="number" value="' . $game['id'] . '" id="gameId" name="gameId">'
-                             . '<h3>'. $game['title'] . '</h3>'
-                             . '<input class="editButton" type="submit" value="Edit">'
-                         . '</form>'
-                         . '<form method="get" action="./deleteItem.php">'
-                            . '<input class="hidden" type="number" value="' . $game['id'] . '" id="gameId" name="gameId">'
-                            . '<input class="deleteButton" type="submit" value="Delete">'
-                         . '</form>'
-                     . '</div>';
+        if (isset($game['title'])) {
+            $return .=     '<div class=editGame>'
+                            . '<form method="get" action="./editEntry.php">'
+                                . '<input class="hidden" type="number" value="' . $game['id'] . '" id="gameId" name="gameId">'
+                                . '<h3>'. $game['title'] . '</h3>'
+                                . '<input class="editButton" type="submit" value="Edit">'
+                            . '</form>'
+                            . '<form method="get" action="./deleteItem.php">'
+                                . '<input class="hidden" type="number" value="' . $game['id'] . '" id="gameId" name="gameId">'
+                                . '<input class="deleteButton" type="submit" value="Delete">'
+                            . '</form>'
+                        . '</div>';
+        }
     }
 
     return $return;
